# Conflicts:
#	src/main/java/com/gofobao/framework/comment/repository/TopicCommentRepository.java
#	src/main/java/com/gofobao/framework/comment/repository/TopicReplyRepository.java
#	src/main/java/com/gofobao/framework/comment/repository/TopicRepository.java
#	src/main/java/com/gofobao/framework/comment/service/TopicReplyService.java
#	src/main/java/com/gofobao/framework/comment/service/impl/TopicCommentServiceImpl.java

+++ src/main/java/com/gofobao/framework/comment/vo/response/VoTopicMemberCenterResp.java
@@ -21,8 +21,8 @@ public class VoTopicMemberCenterResp extends VoBaseResp {
     private String avatar = "";
 
     @ApiModelProperty("帖子最新动态")
-    private Integer topicNumber = 0;
+    private long topicNumber = 0;
 
     @ApiModelProperty("回复我的帖子")
-    private Integer topicReplyNumber = 0;
+    private long topicReplyNumber = 0;
 }
